# docs/openapi.yaml
# Especificacion OpenAPI del Servicio de Donaciones.
# Generado desde la aplicacion (/v3/api-docs) para que no se desincronice del codigo.
# Swagger UI: https://donaciones-xouj.onrender.com/swagger-ui/index.html
#
# Endpoints no modificables por consigna: POST /donaciones, GET /donaciones,
# GET /donaciones/{id}. Se mantienen con ruta y body originales.

openapi: "3.1.0"
info:
  title: "DonaTrack - Servicio de Donaciones"
  version: "1.0.0"
  description: "API del componente de Donaciones. Los endpoints POST/GET /donaciones y GET /donaciones/{id} son no modificables por consigna."
servers:
  - url: "https://donaciones-xouj.onrender.com"
    description: Render
  - url: "http://localhost:8080"
    description: Local
paths:
  "/productos/{id}":
    get:
      tags:
        - productos-controller
      operationId: buscarProductoPorID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ProductoResponse"
    put:
      tags:
        - productos-controller
      operationId: modificarProducto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductoRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ProductoResponse"
    delete:
      tags:
        - productos-controller
      operationId: eliminarProducto
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  "/identificadores/{id}":
    get:
      tags:
        - identificadores-controller
      operationId: buscarIdentificadorPorID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/IdentificadorDTO"
    put:
      tags:
        - identificadores-controller
      operationId: modificarIdentificador
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/IdentificadorDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/IdentificadorDTO"
    delete:
      tags:
        - identificadores-controller
      operationId: eliminarIdentificador
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  "/categorias/{id}":
    get:
      tags:
        - categorias-controller
      operationId: buscarCategoriaPorID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CategoriaResponse"
    put:
      tags:
        - categorias-controller
      operationId: modificarCategoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CategoriaRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CategoriaResponse"
    delete:
      tags:
        - categorias-controller
      operationId: eliminarCategoria
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
  "/productos":
    get:
      tags:
        - productos-controller
      operationId: listarProductos
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/ProductoDTO"
    post:
      tags:
        - productos-controller
      operationId: agregarProducto
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/ProductoRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/ProductoResponse"
  "/identificadores":
    get:
      tags:
        - identificadores-controller
      operationId: listarIdentificadores
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/IdentificadorDTO"
    post:
      tags:
        - identificadores-controller
      operationId: agregarIdentificador
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/IdentificadorDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/IdentificadorDTO"
  "/donaciones":
    get:
      tags:
        - donaciones-controller
      operationId: listarOBuscarPorDonadorYFechaInicio
      parameters:
        - name: donadorID
          in: query
          required: false
          schema:
            type: string
        - name: fecha
          in: query
          required: false
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/DonacionDTO"
    post:
      tags:
        - donaciones-controller
      operationId: registrarDonacion
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/DonacionDTO"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DonacionDTO"
  "/donaciones/{id}/quejas":
    post:
      tags:
        - donaciones-controller
      operationId: registrarQueja
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/QuejaRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DonacionDTO"
  "/donaciones/{id}/queja":
    post:
      tags:
        - donaciones-controller
      operationId: registrarQuejaCompatible
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/QuejaRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DonacionDTO"
  "/categorias":
    get:
      tags:
        - categorias-controller
      operationId: listarCategorias
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CategoriaResponse"
    post:
      tags:
        - categorias-controller
      operationId: agregarCategoria
      requestBody:
        content:
          application/json:
            schema:
              "$ref": "#/components/schemas/CategoriaRequest"
        required: true
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/CategoriaResponse"
  "/donaciones/{id}/estado":
    patch:
      tags:
        - donaciones-controller
      operationId: cambiarEstado
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: estado
          in: query
          required: true
          schema:
            type: string
            enum:
              - INGRESADA
              - ACEPTADA
              - CONQUEJA
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DonacionDTO"
  "/donaciones/{id}":
    get:
      tags:
        - donaciones-controller
      operationId: buscarDonacionPorID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                "$ref": "#/components/schemas/DonacionDTO"
  "/donaciones/{id}/historial":
    get:
      tags:
        - donaciones-controller
      operationId: historialDeEstados
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CambioEstadoResponse"
  "/donaciones/search":
    get:
      tags:
        - donaciones-controller
      operationId: buscarPorDonadorYFechaInicio
      parameters:
        - name: donadorID
          in: query
          required: true
          schema:
            type: string
        - name: fecha
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/DonacionDTO"
  "/categorias/{id}/subcategorias":
    get:
      tags:
        - categorias-controller
      operationId: listarSubcategorias
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: array
                items:
                  "$ref": "#/components/schemas/CategoriaResponse"
  "/admin/estado":
    get:
      tags:
        - admin-controller
      operationId: estado
      responses:
        "200":
          description: OK
          content:
            "*/*":
              schema:
                type: object
                additionalProperties: {}
  "/admin/datos":
    delete:
      tags:
        - admin-controller
      operationId: limpiarDatos
      responses:
        "200":
          description: OK
components:
  schemas:
    ProductoRequest:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        categoriaID:
          type: string
        subcategoriaID:
          type: string
        identificadorID:
          type: string
    ProductoResponse:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        categoriaID:
          type: string
        subcategoriaID:
          type: string
        identificadorID:
          type: string
    IdentificadorDTO:
      type: object
      properties:
        id:
          type: string
        tipo:
          type: string
          enum:
            - QR
            - CODIGODEBARRAS
        descripcion:
          type: string
    CategoriaRequest:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        categoriaPadreID:
          type: string
        subcategoriaID:
          type: string
    CategoriaResponse:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        categoriaPadreID:
          type: string
        esSubcategoria:
          type: boolean
    DonacionDTO:
      type: object
      properties:
        id:
          type: string
        donadorID:
          type: string
        depositoID:
          type: string
        descripcion:
          type: string
        productoID:
          type: string
        cantidad:
          type: integer
          format: int32
        estado:
          type: string
          enum:
            - INGRESADA
            - ACEPTADA
            - CONQUEJA
    QuejaRequest:
      type: object
      properties:
        descripcion:
          type: string
    ProductoDTO:
      type: object
      properties:
        id:
          type: string
        nombre:
          type: string
        descripcion:
          type: string
        categoriaID:
          type: string
        identificadorID:
          type: string
    CambioEstadoResponse:
      type: object
      properties:
        estado:
          type: string
        fechaCambio:
          type: string
          format: date-time
        detalle:
          type: string